Rational<int> -> Rational<int64_t>

+++ include/math/bounds.h
@@ -8,30 +8,30 @@ namespace ir {
 
 class UpperBound {
     Expr expr_;
-    LinearExpr<Rational<int>> lin_;
+    LinearExpr<Rational<int64_t>> lin_;
 
   public:
     UpperBound(const Expr &expr)
         : expr_(expr), lin_{{{getHash(expr), {1, deepCopy(expr)}}}, 0} {}
-    UpperBound(const LinearExpr<Rational<int>> &lin) : lin_(lin) {}
-    UpperBound(LinearExpr<Rational<int>> &&lin) : lin_(std::move(lin)) {}
+    UpperBound(const LinearExpr<Rational<int64_t>> &lin) : lin_(lin) {}
+    UpperBound(LinearExpr<Rational<int64_t>> &&lin) : lin_(std::move(lin)) {}
 
     const Expr &expr();
-    const LinearExpr<Rational<int>> &lin() const { return lin_; }
+    const LinearExpr<Rational<int64_t>> &lin() const { return lin_; }
 };
 
 class LowerBound {
     Expr expr_;
-    LinearExpr<Rational<int>> lin_;
+    LinearExpr<Rational<int64_t>> lin_;
 
   public:
     LowerBound(const Expr &expr)
         : expr_(expr), lin_{{{getHash(expr), {1, expr}}}, 0} {}
-    LowerBound(const LinearExpr<Rational<int>> &lin) : lin_(lin) {}
-    LowerBound(LinearExpr<Rational<int>> &&lin) : lin_(std::move(lin)) {}
+    LowerBound(const LinearExpr<Rational<int64_t>> &lin) : lin_(lin) {}
+    LowerBound(LinearExpr<Rational<int64_t>> &&lin) : lin_(std::move(lin)) {}
 
     const Expr &expr();
-    const LinearExpr<Rational<int>> &lin() const { return lin_; }
+    const LinearExpr<Rational<int64_t>> &lin() const { return lin_; }
 };
 
 UpperBound add(const UpperBound &b1, const UpperBound &b2);
